Refactor: Simplify GUIDs, getGUID()
A broad GUID class should just be 16 random bytes, not effectively a "variant" between a string (player name) and/or the bytes (which may or may not be initialized).

Following the UUIDv4 spec half-consistently is overly complex for no benefit. Just dump the 16 bytes as 32 hex characters, and simplify the storage to an array of 16 bytes.

Generating random bytes in C++ is nasty. Just generate two 64-bit numbers and turn them into bytes. This should be safe.

Entities get their own "EntityGUID" struct which cleanly encapsulated stores the raw bytes + the hex string. We could also generate the hex every time and return a copy from getGUID(), but it doesn't really matter.